improve sync performance

+++ src/main/java/org/craftercms/studio/impl/v2/service/cluster/StudioNodeSyncPublishedTask.java
@@ -28,8 +28,10 @@ import org.craftercms.commons.crypto.TextEncryptor;
 import org.craftercms.commons.crypto.impl.PbkAesTextEncryptor;
 import org.craftercms.studio.api.v1.constant.GitRepositories;
 import org.craftercms.studio.api.v1.dal.RemoteRepository;
+import org.craftercms.studio.api.v1.dal.SiteFeed;
 import org.craftercms.studio.api.v1.deployment.PreviewDeployer;
 import org.craftercms.studio.api.v1.exception.ServiceLayerException;
+import org.craftercms.studio.api.v1.exception.SiteNotFoundException;
 import org.craftercms.studio.api.v1.exception.repository.InvalidRemoteRepositoryCredentialsException;
 import org.craftercms.studio.api.v1.exception.repository.InvalidRemoteRepositoryException;
 import org.craftercms.studio.api.v1.exception.repository.InvalidRemoteUrlException;
@@ -101,7 +103,7 @@ public class StudioNodeSyncPublishedTask implements Runnable {
 
     protected static final Map<String, ReentrantLock> singleWorkerLockMap = new HashMap<String, ReentrantLock>();
     protected static final List<String> createdSites = new ArrayList<String>();
-    protected static final Map<String, List<String>> remotesMap = new HashMap<String, List<String>>();
+    protected static final Map<String, Map<String, String>> remotesMap = new HashMap<String, Map<String, String>>();
 
 
     protected String siteId;
@@ -194,19 +196,34 @@ public class StudioNodeSyncPublishedTask implements Runnable {
                     }
                 }
 
+                boolean syncRequired = true;
+                String siteLastCommitId = StringUtils.EMPTY;
                 try {
-                    logger.debug("Add remotes for site " + siteId);
-                    addRemotes();
-
-                } catch (InvalidRemoteUrlException | ServiceLayerException | CryptoException e) {
-                    logger.error("Error while adding remotes on cluster node for site " + siteId);
+                    SiteFeed siteFeed = siteService.getSite(siteId);
+                    siteLastCommitId = siteFeed.getLastCommitId();
+                    String lastCommitIdRepo = contentRepository.getRepoLastCommitId(siteId);
+                    if (StringUtils.equals(lastCommitIdRepo, siteLastCommitId)) {
+                        syncRequired = false;
+                    }
+                } catch (SiteNotFoundException e) {
+                    logger.error("Site " + siteId + " not found in the database");
                 }
 
-                try {
-                    logger.debug("Update publised repo for site " + siteId);
-                    updatePublished();
-                } catch (IOException | CryptoException | ServiceLayerException e) {
-                    logger.error("Error while updating content for site " + siteId + " on cluster node.", e);
+                if (syncRequired && StringUtils.isNotEmpty(siteLastCommitId)) {
+                    try {
+                        logger.debug("Add remotes for site " + siteId);
+                        addRemotes();
+
+                    } catch (InvalidRemoteUrlException | ServiceLayerException | CryptoException e) {
+                        logger.error("Error while adding remotes on cluster node for site " + siteId);
+                    }
+
+                    try {
+                        logger.debug("Update publised repo for site " + siteId);
+                        updatePublished(siteLastCommitId);
+                    } catch (IOException | CryptoException | ServiceLayerException e) {
+                        logger.error("Error while updating content for site " + siteId + " on cluster node.", e);
+                    }
                 }
             } finally {
                 singleWorkerLock.unlock();
@@ -434,12 +451,12 @@ public class StudioNodeSyncPublishedTask implements Runnable {
     }
 
     private void addRemotes() throws InvalidRemoteUrlException, ServiceLayerException, CryptoException {
-        List<String> existingRemotes = remotesMap.get(siteId);
+        Map<String, String> existingRemotes = remotesMap.get(siteId);
         TextEncryptor encryptor = new PbkAesTextEncryptor(studioConfiguration.getProperty(SECURITY_CIPHER_KEY),
                 studioConfiguration.getProperty(SECURITY_CIPHER_SALT));
         logger.debug("Add cluster members as remotes to local repository");
         for (ClusterMember member : clusterNodes) {
-            if (CollectionUtils.isNotEmpty(existingRemotes) && existingRemotes.contains(member.getGitRemoteName())) {
+            if (existingRemotes != null && existingRemotes.containsKey(member.getGitRemoteName())) {
                 continue;
             }
             FileRepositoryBuilder builder = new FileRepositoryBuilder();
@@ -460,7 +477,7 @@ public class StudioNodeSyncPublishedTask implements Runnable {
                     Set<String> remotes = storedConfig.getSubsections(CONFIG_SECTION_REMOTE);
 
                     if (existingRemotes == null) {
-                        existingRemotes = new ArrayList<String>();
+                        existingRemotes = new HashMap<String, String>();
                         remotesMap.put(siteId, existingRemotes);
                     }
                     if (remotes.contains(member.getGitRemoteName())) {
@@ -483,7 +500,7 @@ public class StudioNodeSyncPublishedTask implements Runnable {
                         remoteAddCommand.setUri(new URIish(remoteUrl));
                         remoteAddCommand.call();
                     }
-                    existingRemotes.add(member.getGitRemoteName());
+                    existingRemotes.put(member.getGitRemoteName(), StringUtils.EMPTY);
                 } catch (URISyntaxException e) {
                     logger.error("Remote URL is invalid " + remoteUrl, e);
                     throw new InvalidRemoteUrlException();
@@ -499,7 +516,7 @@ public class StudioNodeSyncPublishedTask implements Runnable {
         }
     }
 
-    private void updatePublished() throws IOException, CryptoException, ServiceLayerException {
+    private void updatePublished(String lastCommitId) throws IOException, CryptoException, ServiceLayerException {
         logger.debug("Update published repo for site " + siteId);
         boolean toRet = true;
         Path siteSandboxPath = buildRepoPath(PUBLISHED).resolve(GIT_ROOT);
@@ -509,22 +526,33 @@ public class StudioNodeSyncPublishedTask implements Runnable {
                 .readEnvironment()
                 .findGitDir()
                 .build();
+
+        Map<String, String> remoteLastSyncCommits = remotesMap.get(siteId);
+        if (remoteLastSyncCommits == null || remoteLastSyncCommits.isEmpty()) {
+            remoteLastSyncCommits = new HashMap<String, String>();
+            remotesMap.put(siteId, remoteLastSyncCommits);
+        }
         try (Git git = new Git(repo)) {
 
             Set<String> environments = getAllPublishingEnvironments(siteId);
             logger.debug("Update published repo from all active cluster members");
             for (ClusterMember remoteNode : clusterNodes) {
-                logger.debug("Fetch from cluster member " + remoteNode.getLocalIp());
-                final Path tempKey = Files.createTempFile(UUID.randomUUID().toString(), ".tmp");
-                FetchCommand fetch = git.fetch().setRemote(remoteNode.getGitRemoteName());
-                fetch = setAuthenticationForCommand(remoteNode, fetch, tempKey);
-                fetch.call();
-                Files.delete(tempKey);
+                String remoteLastSyncCommit = remoteLastSyncCommits.get(remoteNode.getGitRemoteName());
+                if (StringUtils.isEmpty(remoteLastSyncCommit) ||
+                        !StringUtils.equals(lastCommitId, remoteLastSyncCommit)) {
+                    logger.debug("Fetch from cluster member " + remoteNode.getLocalIp());
+                    final Path tempKey = Files.createTempFile(UUID.randomUUID().toString(), ".tmp");
+                    FetchCommand fetch = git.fetch().setRemote(remoteNode.getGitRemoteName());
+                    fetch = setAuthenticationForCommand(remoteNode, fetch, tempKey);
+                    fetch.call();
+                    Files.delete(tempKey);
 
-                logger.debug("Update all environments for site " + siteId + " from cluster member " +
-                        remoteNode.getLocalIp());
-                for (String branch : environments) {
-                    updatePublishedBranch(git, remoteNode, branch);
+                    logger.debug("Update all environments for site " + siteId + " from cluster member " +
+                            remoteNode.getLocalIp());
+                    for (String branch : environments) {
+                        updatePublishedBranch(git, remoteNode, branch);
+                    }
+                    remoteLastSyncCommits.put(remoteNode.getGitRemoteName(), lastCommitId);
                 }
             }
         } catch (GitAPIException e) {
